+++ src/test/java/org/broadinstitute/hellbender/utils/test/ArgumentsBuilderTest.java
@@ -13,12 +13,12 @@ public final class ArgumentsBuilderTest{
         args.add("--value1");
         args.add(1);
         args.add("Input=")
-            .add(new File("/path/to/somewhere"))
+            .add(new File("path/to/somewhere"))
             .add("Value2=2")
             .add(" Value3= 3 Value4=4");
 
         Assert.assertEquals(args.getArgsArray(), new String[]{"--value1", "1",
-                "--Input", "/path/to/somewhere","--Value2","2","--Value3","3","--Value4","4"});
+                "--Input", "path/to/somewhere","--Value2","2","--Value3","3","--Value4","4"});
     }
     @Test
     public void testOneBigString(){
@@ -31,8 +31,8 @@ public final class ArgumentsBuilderTest{
 
     @Test
     public void testFromArray(){
-        ArgumentsBuilder args = new ArgumentsBuilder(new Object[]{"Option=" + new File("/path/to"), "OtherOption=" + -1, new File("/somewhere")});
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--Option","/path/to", "--OtherOption", "-1", "/somewhere"});
+        ArgumentsBuilder args = new ArgumentsBuilder(new Object[]{"Option=" + new File("path/to"), "OtherOption=" + -1, new File("somewhere")});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--Option","path/to", "--OtherOption", "-1", "somewhere"});
     }
 
     @Test
@@ -62,41 +62,41 @@ public final class ArgumentsBuilderTest{
     @Test
     public void testAddFileArgument() throws Exception {
         ArgumentsBuilder args = new ArgumentsBuilder();
-        File file = new File("/path/to/somewhere");
+        File file = new File("path/to/somewhere");
         args.addFileArgument("foo", file);
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--foo", file.toString()});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--foo", file.getAbsolutePath()});
     }
 
     @Test
     public void testInput() throws Exception {
         ArgumentsBuilder args = new ArgumentsBuilder();
-        File file = new File("/path/to/somewhere");
+        File file = new File("path/to/somewhere");
         args.addInput(file);
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.INPUT_LONG_NAME, file.toString()});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.INPUT_LONG_NAME, file.getAbsolutePath()});
     }
 
     @Test
     public void testOutput() throws Exception {
         ArgumentsBuilder args = new ArgumentsBuilder();
-        File file = new File("/path/to/somewhere");
+        File file = new File("path/to/somewhere");
         args.addOutput(file);
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.OUTPUT_LONG_NAME, file.toString()});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.OUTPUT_LONG_NAME, file.getAbsolutePath()});
     }
 
     @Test
     public void testReference() throws Exception {
         ArgumentsBuilder args = new ArgumentsBuilder();
-        File file = new File("/path/to/somewhere");
+        File file = new File("path/to/somewhere");
         args.addReference(file);
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.REFERENCE_LONG_NAME, file.toString()});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.REFERENCE_LONG_NAME, file.getAbsolutePath()});
     }
 
     @Test
     public void testVcf() throws Exception {
         ArgumentsBuilder args = new ArgumentsBuilder();
-        File file = new File("/path/to/somewhere");
+        File file = new File("path/to/somewhere");
         args.addVCF(file);
-        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.VARIANT_LONG_NAME, file.toString()});
+        Assert.assertEquals(args.getArgsArray(), new String[]{"--" + StandardArgumentDefinitions.VARIANT_LONG_NAME, file.getAbsolutePath()});
     }
 
 }
